main.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# This is a placeholder for running the ADK web UI.
# In a real scenario, you would typically run this from the command line
# using `adk web` from the directory containing your agent.
# For demonstration purposes, we'll just indicate that the setup is complete.

print("ADK agent setup complete. To run the ADK web UI, navigate to this directory in your terminal (with the virtual environment activated) and run: adk web")
print("You will also need to set up Google Cloud credentials and enable the Vertex AI API.")